Add files into modules
